Changed the sharedReport query, now we can query for limited or unlimited or both types of shareReports. Also now deleting report will throw an error if your report is shared with a doctor. You can also force delete reports. Also added some preauthorization only basicUser and patients can add share reports

+++ backend/pinklifeline/src/main/java/com/sadi/pinklifeline/service/ReportHandlerService.java
@@ -83,8 +83,15 @@ public class ReportHandlerService {
     }
 
     @PreAuthorize("hasAnyRole('BASICUSER', 'PATIENT')")
-    public void deleteReport(Long id) {
+    public void deleteReport(Long id, Boolean force) {
         Long userId = SecurityUtils.getOwnerID();
+        List<ReportSharedInfoDTO> dtos = sharedReportRepository.findSharedReportInfoByReportId(id,
+                LocalDateTime.now());
+        if(!dtos.isEmpty() && !force){
+            throw new BadRequestFromUserException(
+                    String.format("Report with id:%d is shared to %d doctors", id, dtos.size())
+            );
+        }
         Report report = getReport(id);
         verifyReportOwner(report, userId);
         reportRepository.delete(report);
